refactor(api-journeys): rename event/utils.ts to eventService.ts, hide test injectors (ENG-3725 A31)
The file is the event-write service, not generic utils. The
__set*QueueForTests injectors and the lazy queue holders move to an
internal event/queues.ts module so they no longer sit on the service
surface the 17 production importers see.
