+++ apis/api-journeys/src/schema/event/button/buttonClickEventCreate.mutation.ts
@@ -1,7 +1,7 @@
 import { Prisma, prisma } from '@core/prisma/journeys/client'
 
 import { builder } from '../../builder'
-import { appendEventToGoogleSheets, validateBlockEvent } from '../utils'
+import { appendEventToGoogleSheets, validateBlockEvent } from '../eventService'
 
 import { ButtonClickEventRef } from './buttonClickEvent'
 import { ButtonClickEventCreateInput } from './inputs'
