+++ apis/api-journeys/src/schema/event/multiselect/multiselectSubmissionEventCreate.mutation.ts
@@ -3,7 +3,7 @@ import { v4 as uuidv4 } from 'uuid'
 import { prisma } from '@core/prisma/journeys/client'
 
 import { builder } from '../../builder'
-import { appendEventToGoogleSheets, validateBlockEvent } from '../utils'
+import { appendEventToGoogleSheets, validateBlockEvent } from '../eventService'
 
 import { MultiselectSubmissionEventCreateInput } from './inputs/multiselectSubmissionEventCreateInput'
 import { MultiselectSubmissionEventRef } from './multiselectSubmissionEvent'
