+++ apis/api-journeys/src/schema/event/step/stepPreviousEventCreate.mutation.ts
@@ -1,7 +1,7 @@
 import { prisma } from '@core/prisma/journeys/client'
 
 import { builder } from '../../builder'
-import { validateBlockEvent } from '../utils'
+import { validateBlockEvent } from '../eventService'
 
 import { StepPreviousEventCreateInput } from './inputs'
 import { StepPreviousEventRef } from './stepPreviousEvent'
