+++ apis/api-journeys/src/schema/event/video/videoExpandEventCreate.mutation.ts
@@ -1,7 +1,7 @@
 import { prisma } from '@core/prisma/journeys/client'
 
 import { builder } from '../../builder'
-import { validateBlockEvent } from '../utils'
+import { validateBlockEvent } from '../eventService'
 
 import { VideoExpandEventCreateInput } from './inputs'
 import { VideoExpandEventRef } from './videoExpandEvent'
