+++ apis/api-journeys/src/schema/event/video/videoPlayEventCreate.mutation.ts
@@ -1,7 +1,7 @@
 import { prisma } from '@core/prisma/journeys/client'
 
 import { builder } from '../../builder'
-import { resetEventsEmailDelay, validateBlockEvent } from '../utils'
+import { resetEventsEmailDelay, validateBlockEvent } from '../eventService'
 
 import { VideoPlayEventCreateInput } from './inputs'
 import { VideoPlayEventRef } from './videoPlayEvent'
